% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{quire_error}
\alias{quire_error}
\title{A block the host could not draw}
\usage{
quire_error(message)
}
\arguments{
\item{message}{Why, for the reader.}
}
\value{
An error drawing, shown in the block's place.
}
\description{
A block the host could not draw
}
